fix(plan): fix send/pull operations failure on armv7 architectures
Before this fix, an implicit conversion from size_t to uint64_t used to
occur during a send. This is harmless on x86_64 because the two types
are the same but not on armv7. This cast was causing pull tests to fail
on the armv7 architecture.
This commit fix this issue.

+++ include/synctl/plan/Send_1.hxx
@@ -2,6 +2,7 @@
 #define _INCLUDE_SYNCTL_SEND_1_HXX_
 
 
+#include <cstdint>
 #include <string>
 
 #include "synctl/tree/Filter.hxx"
@@ -36,10 +37,10 @@ class Send_1
 	void _sendObject(const Context *context);
 
 	void _transfer(const Context *context, InputStream *input,
-		       size_t size);
+		       uint64_t size);
 
 	void _sendDirectory(const Context *context, InputStream *input,
-			    size_t size);
+			    uint64_t size);
 
 
  public:
